RPi[x]: adding dhcp command, modify boot command
For booting a kernel via DHCP the dhcp command is expanded to load the
provided kernel into ${kernel_addr_r} and starting the kernel after
loading.

The bootcmd is modified to check for availability of a file uEnv.txt on
the mmc0 device so the user can add U-Boot commands without the need to
create a boot.scr or boot.scr.uimg.

Patch based on
  https://raspberrypi.org/forums/viewtopic.php?f=29&t=109292